[#112] feat: LiveWritingView 작성 필드 UI 간략 작성
