+++ lib/features/live_writing/presentation/model/live_writing_state.dart
@@ -1,0 +1,10 @@
+enum LiveWritingState {
+  ready,
+  writing,
+  done,
+}
+
+enum LiveBubbleState {
+  showingDefault,
+  showingDetail,
+}
